Arreglada configuración WIFI

+++ Extension_WIFI_ESP8266_and_IOT_mrtduino/extension/src/IOT_ts.h
@@ -15,7 +15,7 @@ class Iot_device
     bool thingSpeakWrite(unsigned int,float ,bool );
 	bool iftttWrite(String,float value1,float value2,float value3,bool);
 	void Reset_ESP8266(int ,bool );
-    void Baudios_ESP8266(int,int,bool);
+    void Baudios_ESP8266(long int,int,bool);
     void Mode_ESP8266(unsigned int ,int ,bool);
 	void Credentials_ESP8266(String ,String,int,bool);
 	private:
